+++ src/components/Editor.tsx
@@ -3,14 +3,14 @@ import React, { useEffect, useState } from "react";
 import { LoadingScreen } from "../internal/puck/components/LoadingScreen.tsx";
 import { Toaster } from "../internal/puck/ui/Toaster.tsx";
 import { type Config } from "@measured/puck";
-import { useMessageSenders } from "../internal/hooks/useMessage.ts";
 import "@measured/puck/puck.css";
 import { DevLogger } from "../utils/devLogger.ts";
 import { ThemeConfig } from "../utils/themeResolver.ts";
 import { useQuickFindShortcut } from "../internal/hooks/useQuickFindShortcut.ts";
-import { useMessageReceivers } from "../hooks/useMessageReceivers.ts";
+import { useMessageReceivers } from "../internal/hooks/useMessageReceivers.ts";
 import { LayoutEditor } from "../internal/components/LayoutEditor.tsx";
 import { ThemeEditor } from "../internal/components/ThemeEditor.tsx";
+import { useCommonMessageSenders } from "../internal/hooks/useMessageSenders.ts";
 
 const devLogger = new DevLogger();
 
@@ -46,7 +46,7 @@ export const Editor = ({
     themeSaveStateFetched,
   } = useMessageReceivers(componentRegistry);
 
-  const { pushPageSets } = useMessageSenders();
+  const { pushPageSets } = useCommonMessageSenders();
 
   useQuickFindShortcut();
 
@@ -104,14 +104,14 @@ export const Editor = ({
     !(themeSaveStateFetched || !templateMetadata?.isThemeMode);
 
   const progress: number =
-    90 * // @ts-expect-error adding bools is fine
+    100 * // @ts-expect-error adding bools is fine
     ((!!puckConfig +
       !!templateMetadata +
       !!document +
-      (layoutSaveStateFetched || templateMetadata?.isThemeMode) +
+      ((layoutSaveStateFetched || templateMetadata?.isThemeMode) ?? 0) +
       visualConfigurationDataFetched +
-      (themeDataFetched || !templateMetadata?.isThemeMode) +
-      (themeSaveStateFetched || !templateMetadata?.isThemeMode)) /
+      ((themeDataFetched || !templateMetadata?.isThemeMode) ?? 0) +
+      ((themeSaveStateFetched || !templateMetadata?.isThemeMode) ?? 0)) /
       7);
 
   return (
