+++ src/components/index.ts
@@ -1,7 +1,6 @@
 export { Editor } from "./Editor.tsx";
 export { EntityField } from "./EntityField.tsx";
 export { YextEntityFieldSelector } from "./YextEntityFieldSelector.tsx";
-export { ColorSelector } from "./ColorSelector.tsx";
 export type {
   RenderYextEntityFieldSelectorProps,
   YextEntityField,
