+++ src/internal/hooks/layout/useLocalStorage.ts
@@ -1,26 +1,24 @@
 import { useCallback } from "react";
-import { TemplateMetadata } from "../types/templateMetadata.ts";
-import { DevLogger } from "../../utils/devLogger.ts";
+import { TemplateMetadata } from "../../types/templateMetadata.ts";
+import { DevLogger } from "../../../utils/devLogger.ts";
 
-export const Role = {
+const Roles = {
   GLOBAL: "global",
   INDIVIDUAL: "individual",
 };
 
 const ROLE = "ROLE_",
-  SITE = "SITE_",
   TEMPLATE = "TEMPLATE_",
   LAYOUT = "LAYOUT_",
-  ENTITY = "ENTITY_",
-  THEME = "THEME_";
+  ENTITY = "ENTITY_";
 
 const devLogger = new DevLogger();
 
 /**
- * useLocalStorage contains helper functions for reading and clearing local storage
+ * useLayoutLocalStorage contains helper functions for reading and clearing local storage
  * @param templateMetadata The template metadata for the active template (undefined if not loaded)
  */
-export const useLocalStorage = (
+export const useLayoutLocalStorage = (
   templateMetadata: TemplateMetadata | undefined
 ) => {
   const buildVisualConfigLocalStorageKey = useCallback(() => {
@@ -37,26 +35,6 @@ export const useLocalStorage = (
     );
   }, [templateMetadata]);
 
-  const buildThemeLocalStorageKey = useCallback(() => {
-    if (!templateMetadata) {
-      return "";
-    }
-
-    return getThemeLocalStorageKey(
-      templateMetadata.isDevMode,
-      templateMetadata.siteId,
-      templateMetadata.themeEntityId
-    );
-  }, [templateMetadata]);
-
-  /**
-   * Clears the user's theming in localStorage
-   */
-  const clearThemeLocalStorage = () => {
-    devLogger.logFunc("clearThemeLocalStorage");
-    window.localStorage.removeItem(buildThemeLocalStorageKey());
-  };
-
   /**
    * Clears the user's visual configuration in localStorage and resets the current Puck history
    */
@@ -66,10 +44,8 @@ export const useLocalStorage = (
   };
 
   return {
-    clearThemeLocalStorage,
     clearVisualConfigLocalStorage,
     buildVisualConfigLocalStorageKey,
-    buildThemeLocalStorageKey,
   };
 };
 
@@ -86,7 +62,7 @@ function getVisualConfigLocalStorageKey(
       `Unable to generate local storage key, missing query parameters${!entityId && " entityId"}${!layoutId && " layoutId"}`
     );
   }
-  if (role === Role.INDIVIDUAL) {
+  if (role === Roles.INDIVIDUAL) {
     if (!entityId) {
       throw new Error(`EntityId required for role ${role}`);
     }
@@ -97,17 +73,4 @@ function getVisualConfigLocalStorageKey(
     throw new Error(`LayoutId required for role ${role}`);
   }
   return devPrefix + ROLE + role + TEMPLATE + templateId + LAYOUT + layoutId;
-}
-
-function getThemeLocalStorageKey(
-  isDevMode: boolean,
-  siteId: number,
-  themeEntityId?: number
-): string {
-  if (!themeEntityId) {
-    throw new Error(
-      "Unable to generate local storage key for themes, missing themeEntityId"
-    );
-  }
-  return (isDevMode ? "dev" : "") + SITE + siteId + THEME + themeEntityId;
 }
