+++ src/internal/hooks/layout/useMessageSenders.ts
@@ -1,0 +1,30 @@
+import { TARGET_ORIGINS, useSendMessageToParent } from "../useMessage.ts";
+
+export const useLayoutMessageSenders = () => {
+  const { sendToParent: saveLayoutSaveState } = useSendMessageToParent(
+    "saveSaveState",
+    TARGET_ORIGINS
+  );
+
+  const { sendToParent: deleteLayoutSaveState } = useSendMessageToParent(
+    "deleteSaveState",
+    TARGET_ORIGINS
+  );
+
+  const { sendToParent: publishVisualConfiguration } = useSendMessageToParent(
+    "saveVisualConfigData",
+    TARGET_ORIGINS
+  );
+
+  const { sendToParent: sendDevLayoutSaveStateData } = useSendMessageToParent(
+    "sendDevSaveStateData",
+    TARGET_ORIGINS
+  );
+
+  return {
+    saveLayoutSaveState,
+    deleteLayoutSaveState,
+    publishVisualConfiguration,
+    sendDevLayoutSaveStateData,
+  };
+};
