+++ src/internal/hooks/useMessage.ts
@@ -226,74 +226,3 @@ const statusSetter = (setStatus: Dispatch<SetStateAction<MessageStatus>>) => {
     }
   };
 };
-
-export const useMessageSenders = () => {
-  const { sendToParent: iFrameLoaded } = useSendMessageToParent(
-    "iFrameLoaded",
-    TARGET_ORIGINS
-  );
-
-  const { sendToParent: sendDevLayoutSaveStateData } = useSendMessageToParent(
-    "sendDevSaveStateData",
-    TARGET_ORIGINS
-  );
-
-  const { sendToParent: pushPageSets } = useSendMessageToParent(
-    "pushPageSets",
-    TARGET_ORIGINS
-  );
-
-  const { sendToParent: saveLayoutSaveState } = useSendMessageToParent(
-    "saveSaveState",
-    TARGET_ORIGINS
-  );
-
-  const { sendToParent: deleteLayoutSaveState } = useSendMessageToParent(
-    "deleteSaveState",
-    TARGET_ORIGINS
-  );
-
-  const { sendToParent: publishVisualConfiguration } = useSendMessageToParent(
-    "saveVisualConfigData",
-    TARGET_ORIGINS
-  );
-
-  const { sendToParent: saveThemeSaveState } = useSendMessageToParent(
-    "saveThemeSaveState",
-    TARGET_ORIGINS
-  );
-
-  const { sendToParent: deleteThemeSaveState } = useSendMessageToParent(
-    "deleteThemeSaveState",
-    TARGET_ORIGINS
-  );
-
-  const { sendToParent: publishThemeConfiguration } = useSendMessageToParent(
-    "saveThemeData",
-    TARGET_ORIGINS
-  );
-
-  const { sendToParent: openQuickFind } = useSendMessageToParent(
-    "openQuickFind",
-    TARGET_ORIGINS
-  );
-
-  const { sendToParent: sendDevThemeSaveStateData } = useSendMessageToParent(
-    "sendDevThemeSaveStateData",
-    TARGET_ORIGINS
-  );
-
-  return {
-    iFrameLoaded,
-    sendDevLayoutSaveStateData,
-    pushPageSets,
-    saveLayoutSaveState,
-    deleteLayoutSaveState,
-    publishVisualConfiguration,
-    saveThemeSaveState,
-    deleteThemeSaveState,
-    publishThemeConfiguration,
-    openQuickFind,
-    sendDevThemeSaveStateData,
-  };
-};
